+++ iters/constants.py
@@ -1,9 +1,9 @@
 __all__ = ("EMPTY_BYTES", "EMPTY_STRING", "DEFAULT_START", "DEFAULT_STEP")
 
 EMPTY_BYTES = bytes()
-"""An empty bytes instance."""
+"""An empty [`bytes`][bytes] instance."""
 EMPTY_STRING = str()
-"""An empty string instance."""
+"""An empty [`str`][str] instance."""
 
 DEFAULT_START = 0
 """The default start value for counting iterators."""