+++ iters/typing.py
@@ -1,11 +1,20 @@
 from __future__ import annotations
 
-from abc import abstractmethod as required
-from typing import Protocol, TypeVar, runtime_checkable
+from typing import Optional, Protocol, TypeVar, runtime_checkable
 
-__all__ = ("Sum", "Product")
+from typing_aliases import Predicate, required
+from typing_extensions import Self
 
-S = TypeVar("S", bound="Sum")
+__all__ = ("OptionalPredicate", "Sum", "Product")
+
+T = TypeVar("T")
+
+OptionalPredicate = Optional[Predicate[T]]
+"""Represents optional predicates.
+
+Passing [`None`][None] is equivalent to passing [`bool`][bool], though most functions
+are optimized to reduce the overhead of calling [`bool`][bool].
+"""
 
 
 @runtime_checkable
@@ -13,11 +22,8 @@ class Sum(Protocol):
     """Represents types for which adding `self: S` to `other: S` returns `S`."""
 
     @required
-    def __add__(self: S, __other: S) -> S:
+    def __add__(self, __other: Self) -> Self:
         raise NotImplementedError
-
-
-P = TypeVar("P", bound="Product")
 
 
 @runtime_checkable
@@ -25,5 +31,5 @@ class Product(Protocol):
     """Represents types for which multiplying `self: P` with `other: P` returns `P`."""
 
     @required
-    def __mul__(self: P, __other: P) -> P:
+    def __mul__(self, __other: Self) -> Self:
         raise NotImplementedError