+++ iters/wraps.py
@@ -1,9 +1,7 @@
 from typing import Iterable, Iterator, TypeVar
 
 from typing_aliases import Binary, Unary
-from wraps.option import Option, is_some
-
-__all__ = ("filter_map_option", "scan")
+from wraps.option import Option
 
 S = TypeVar("S")
 T = TypeVar("T")
@@ -14,7 +12,7 @@ def scan(state: S, function: Binary[S, T, Option[U]], iterable: Iterable[T]) -> 
     for item in iterable:
         option = function(state, item)
 
-        if is_some(option):
+        if option.is_some():
             yield option.unwrap()
 
         else:
@@ -22,6 +20,8 @@ def scan(state: S, function: Binary[S, T, Option[U]], iterable: Iterable[T]) -> 
 
 
 def filter_map_option(function: Unary[T, Option[U]], iterable: Iterable[T]) -> Iterator[U]:
-    for option in map(function, iterable):
-        if is_some(option):
+    for item in iterable:
+        option = function(item)
+
+        if option.is_some():
             yield option.unwrap()